Rename class: GameChatHistoryDao->LobbyGameDao
