+++ http-server/src/main/java/org/triplea/db/dao/lobby/games/LobbyGameDao.java
@@ -1,4 +1,4 @@
-package org.triplea.db.dao.chat.history;
+package org.triplea.db.dao.lobby.games;
 
 import com.google.common.base.Ascii;
 import org.jdbi.v3.sqlobject.customizer.Bind;
@@ -9,7 +9,7 @@ import org.triplea.http.client.lobby.chat.upload.ChatMessageUpload;
  * Game chat history table stores chat messages that have happened in games. This data is upload by
  * game servers to the lobby and is then recorded in database.
  */
-public interface GameChatHistoryDao {
+public interface LobbyGameDao {
   int MESSAGE_COLUMN_LENGTH = 240;
 
   default void recordChat(final ChatMessageUpload chatMessageUpload) {
@@ -22,7 +22,11 @@ public interface GameChatHistoryDao {
 
   @SqlUpdate(
       "insert into game_chat_history (host_name, game_id, username, message) "
-          + "values(:hostname, :gameId, :username, :message)")
+          + "values("
+          + ":hostname, "
+          + " (select id from lobby_game where game_id = :gameId),"
+          + ":username, "
+          + ":message)")
   void insertChatMessage(
       @Bind("hostname") String hostname,
       @Bind("gameId") String gameId,
