+++ http-server/src/test/java/org/triplea/db/dao/lobby/games/LobbyGameDaoTest.java
@@ -1,17 +1,18 @@
-package org.triplea.db.dao.chat.history;
+package org.triplea.db.dao.lobby.games;
 
 import com.github.database.rider.core.api.dataset.DataSet;
 import org.junit.jupiter.api.Test;
 import org.triplea.db.dao.DaoTest;
+import org.triplea.db.dao.lobby.games.LobbyGameDao;
 import org.triplea.http.client.lobby.chat.upload.ChatMessageUpload;
 
-class GameChatHistoryDaoTest extends DaoTest {
-  private final GameChatHistoryDao gameChatHistoryDao = DaoTest.newDao(GameChatHistoryDao.class);
+class LobbyGameDaoTest extends DaoTest {
+  private final LobbyGameDao lobbyGameDao = DaoTest.newDao(LobbyGameDao.class);
 
   @Test
   @DataSet(cleanBefore = true, value = "chat_history/game_chat_history_insert_after.yml")
   void insertChatMessage() {
-    gameChatHistoryDao.recordChat(
+    lobbyGameDao.recordChat(
         ChatMessageUpload.builder()
             .hostName("hostname")
             .gameId("gameid-100")
